+++ app/controllers/metus/pages_controller.rb
@@ -1,30 +1,73 @@
 # frozen_string_literal: true
 
+require 'asciidoctor'
+
 module Metus
   class PagesController < Metus.public_controller_class
-    def self.filename_to_pagename(filename)
+    def self.filename_to_pagename(directory, name)
+      [directory, name].select(&:present?).join('/').to_sym
+    end
+
+    def self.parse_page(filename)
       directory = File.dirname(filename)
       directory.gsub!(/#{PAGES_DIRECTORY}[\/]?/, '')
 
-      name = File.basename(filename, '.html.haml')
+      name, extension = basename(File.basename(filename))
+      pagename = filename_to_pagename(directory, name)
 
-      [directory, name].select(&:present?).join('/').to_sym
+      {
+        name: name,
+        extension: extension,
+        template: template(pagename, extension),
+        directory: directory,
+        pagename: pagename,
+        actionname: path_to_actionname(pagename),
+        filename: filename
+      }
+    end
+
+    def self.basename(filename)
+      [
+        split_extension(filename, '.html.haml'),
+        split_extension(filename, '.adoc')
+      ].compact.first
+    end
+
+    def self.split_extension(filename, extension)
+      name = File.basename(filename, extension)
+
+      if "#{name}#{extension}" == filename
+        return [name, extension]
+      end
     end
 
     def self.path_to_actionname(path)
       path.to_s.gsub('/', '__slash__')
     end
 
+    def self.template(name, extension)
+      case extension
+      when '.html.haml'
+        "pages/#{name}.html.haml"
+      when '.adoc'
+        'metus/pages/asciidoc.html.haml'
+      end
+    end
+
     PAGES_DIRECTORY = Rails.root.join('app/views/pages')
 
-    PAGES = Dir[PAGES_DIRECTORY.join('**/*')]
+    PAGES_WITH_TEMPLATES = Dir[PAGES_DIRECTORY.join('**/*')]
       .reject { |filename| File.directory?(filename) }
-      .map { |filename| Metus::PagesController.filename_to_pagename(filename) }
+      .map { |filename| Metus::PagesController.parse_page(filename) }
+
+    PAGES = PAGES_WITH_TEMPLATES
+      .map { |page| page[:pagename] }
       .map(&:to_sym)
 
-    PAGES.each do |page|
-      define_method(Metus::PagesController.path_to_actionname(page)) do
-        render "pages/#{page}.html.haml"
+    PAGES_WITH_TEMPLATES.each do |page|
+      define_method(page[:actionname]) do
+        @page = page
+        render page[:template]
       end
     end
   end